requirements.txt
Flask==2.2.5
face-recognition==1.3.0
Pillow==8.4.0
opencv-python-headless==4.5.4.60
pandas==1.3.4
Werkzeug==2.2.3
python-dotenv==0.19.0
numpy==1.24.4
Flask-SQLAlchemy==3.0.3
SQLAlchemy==1.4.46
pinecone==5.4.2
psycopg2-binary==2.9.9
firebase-admin